chip: migrate onto the Core 0 input pipeline (#107)
src/engines/chip/midi_controller.cpp -> input_subsystem.cpp, mirroring
subtractive's and fm's now-established pattern exactly: NOTE/MODIFIER/
CONFIGURATION route through input_dispatch() into chip's own
kMappingTable, and midi_controller_process() is a one-line call into
the shared midi_controller_process_generic() (src/midi/midi_dispatch.h,
midi_controller_generic.h).

Voice allocation moved into set_note() (the Voice Allocation
Interface), same fix already applied to subtractive/fm. Filter-bus
binding (bind_filter()/release_bus()) -- chip's own second,
allocator-like concern alongside voice_alloc -- moves into set_note()
too, at the same point in the same Handler, since its lifecycle
already matched voice allocation's timing (resolved once per note-on);
instrument selection (Configuration) never touches it directly, a
channel's binding only changes on its next note-on, unchanged from the
pre-migration behavior.

CC16 (the BeatStep-Pro-safe encoder alternative to Program Change) is
dropped -- module_chip.md's own history documents the identical
rationale fm's CC30 had, and both were duplicate logic once Program
Change routes through CONFIGURATION. Program Change alone now selects
the instrument, per the standing Program-Change-alone convention.

Verified: all six engines build clean, full host test suite passes.
